refactor: share the base agent runtime prompt and remove team mode
- move the shared runtime prompt composition into `agent-runtime-prompt.ts` and keep OpenCode MCP alias guidance in a harness-specific prompt helper
- build harness system prompts from a mandatory base prefix first, then append workspace `AGENTS.md` instructions as subordinate workspace guidance
- strengthen the base prompt wording so the runtime instructions must always be followed and tool, skill, and MCP usage is explicitly prioritized
- remove legacy team-mode parsing and request shapes so workspace plans, runner config requests, and harness-host contracts only support a single agent
- update runtime config, harness conformance, workspace plan, and contract tests to cover the new prompt ordering and single-agent-only behavior
- validation: `npm test` and `npm run typecheck` in `runtime/api-server`; `npm test` and `npm run typecheck` in `runtime/harness-host`
